Creating new Schedule Modal Option
