+++ src/css/Modal.css
@@ -1,0 +1,13 @@
+.modal-head{
+  margin-top: 20%;
+}
+.modal-height{
+  padding-top: 10px;
+
+}
+.modal-header-align{
+  margin-right: auto;
+}
+.modal-header{
+  background  : linear-gradient(#007095,#20C3D3);
+}
